refactor(section): 구간 제거 인수 테스트를 등록인수 테스트와 같은 파일로 모았다.
